requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.0

python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
